[type:refactor] refactor shenyu-logging-common
