+++ shenyu-plugin/shenyu-plugin-logging/shenyu-plugin-logging-kafka/src/test/java/org/apache/shenyu/plugin/logging/kafka/collector/DefaultLogCollectorTest.java
@@ -19,10 +19,9 @@ package org.apache.shenyu.plugin.logging.kafka.collector;
 
 import java.lang.reflect.Field;
 
-import org.apache.shenyu.plugin.logging.common.LogConsumeClient;
+import org.apache.shenyu.plugin.logging.common.client.LogConsumeClient;
 import org.apache.shenyu.plugin.logging.common.collector.AbstractLogCollector;
 import org.apache.shenyu.plugin.logging.common.entity.ShenyuRequestLog;
-import org.apache.shenyu.plugin.logging.kafka.collector.DefaultLogCollector;
 import org.apache.shenyu.plugin.logging.kafka.kafka.KafkaLogCollectClient;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
